adobe_india_hackathon-main/round1a/requirements.txt
PyMuPDF>=1.26.0
pdfplumber>=0.11.0
pdfminer.six>=20250506
pytesseract>=0.3.13
Pillow>=11.0.0
pandas>=2.3.0
numpy>=1.23.0
